Reorganizing provider package
