+++ android-sdk/src/main/java/co/elastic/apm/android/sdk/internal/providers/Provider.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package co.elastic.apm.android.sdk.providers;
+package co.elastic.apm.android.sdk.internal.providers;
 
 public interface Provider<T> {
