+++ android-sdk/src/main/java/co/elastic/apm/android/sdk/traces/http/attributes/visitors/ConnectionHttpAttributes.java
@@ -19,7 +19,7 @@
 package co.elastic.apm.android.sdk.traces.http.attributes.visitors;
 
 import co.elastic.apm.android.sdk.ElasticApmAgent;
-import co.elastic.apm.android.sdk.providers.LazyProvider;
+import co.elastic.apm.android.sdk.internal.providers.LazyProvider;
 import co.elastic.apm.android.sdk.internal.services.Service;
 import co.elastic.apm.android.sdk.internal.services.network.NetworkService;
 import co.elastic.apm.android.sdk.internal.services.network.data.type.NetworkType;
